requirements.txt
Flask==2.3.3
flask-bootstrap==3.3.7.1
Flask-WTF==1.2.1
Web3==6.17.0
cryptography==42.0.5
python-dotenv==1.0.1
WTForms==3.1.2
Werkzeug==2.3.8
email-validator==2.1.1
